Start 10 client threads. [Fixes #4]
Assuming that latency to us-east from australia is about 250ms,
posting an https payload could take up to 3s. With one thread
that's a maximum of 20/minute, with 10 that's 200/minute.

This is signficantly higher than the rate-limit applied inside
the Bugsnag data-center of 60/minute, but because the Bugsnag rate
limit applies only to duplicate instances of the same exception,
it's a good idea to exceed it.

Assuming that error payloads are around 60kb, which is about
normal for those with code inline, this change could push the
bandwidth requirements up from 150kbps to 1.5Mbps.